Getting Circular
Changed the jquery post call triggered from the user's drag event to call the 'opponentsTurn' function which takes an array of the player names. The function is called recursively shifting off an opponents name each time until it's empty at which point the user is informed it's his/her turn again.